+++ logging-utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "google-logging-tools",
+  "name": "google-logging-utils",
   "version": "0.0.0",
   "description": "An ad-hoc debug logger package for other libraries",
   "main": "build/src/index.js",
