+++ logging-utils/samples/typescript/quickstart.ts
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     https://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,4 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export * from './logging-tools';
+// sample-metadata:
+//   title: Quickstart
+//   description: A quick introduction to using the Pub/Sub client library.
+//   usage: node quickstart.js <project-id> <topic-name-or-id> <subscription-name-or-id>
+
+// [START adhoc_logging_quickstart]
+import {log} from 'google-logging-utils';
+
+function main() {
+  const test = log('testing');
+  test({other: {foo: 'bar'}}, 'boo');
+  test.info('info');
+}
+// [END adhoc_logging_quickstart]
+
+main();
