+++ logging-utils/test/logging-utils.ts
@@ -15,7 +15,7 @@
 import {describe, it} from 'mocha';
 import * as assert from 'assert';
 
-import * as al from '../src/logging-tools';
+import * as al from '../src/logging-utils';
 
 interface TestLog {
   namespace: string;
