+++ release-please-config.json
@@ -9,7 +9,7 @@
   "packages": {
     "gax": {},
     "tools": {},
-    "logging-tools": {}
+    "logging-utils": {}
   },
   "$schema": "https://raw.githubusercontent.com/googleapis/release-please/main/schemas/config.json"
 }
